remove duplicate old keys folder
