+++ apps/dashboard/src/app/app/keys/_actions/create-api-key.ts
@@ -14,7 +14,7 @@ export const createApiKeyAction = authorizedAction
     })
   )
   .action(async ({ parsedInput: { description: description } }) => {
-    const apiKey = await api.apiKeys.create({
+    const apiKey = await api.hippoKeys.create({
       description,
     });
